docuent scripts, and remove "0" from issue directory
Hopefully some comments help explain what's going on. This pass also
caught more TODOs that we out of date.

The DVDs use a two-digit, zero-padded issue number in their directory
naming. No year on disc 1 has more than 9 issues, so it's not necessary.
If there were ten or more issues in a year, the code as written would have
produced an unhelpful _010 anyway.