html-escape alt-text unconditionally
Issue 75 in 1999 has OCR text of "&amp;afi;". No idea why, because
comparing the surrounding text to the page image suggests it was reading
"1-800-" at that position.

Regardless, xq unencodes the "&amp;" to leave "&afi;". Liquid's
`encode_once` leaves the "&afi;" in place, even though that's an invalid
character entity. So, changing to `encode` to get us back to "&amp;afi;".