actually use the issue-derived date in the EPUB
I've been trying to keep the estimation of the original publication date
available, and only just realized that I stopped using it back when I
switched from generating a website to generating EPUB.

It turns out that EPUB has no official place to record the date of the
original publication that has been digitized by the document. The
`dc:date` metadata property has been formally specified in EPUB 3 to be
for the date of the EPUB document's creation. I've followed a tip to stick
it in the generic "meta" properties as a "dcterms:issued" that refines the
"dc:source" (which I had set to the hash-looking "docid" string I found in
the source document.xml).